+++ examples/node/createObject.js
@@ -12,12 +12,15 @@ const client = new fsm.CoreAPIClient({
   clientVersion: packageJson.version,
 
   // put your auth config here
-  authAccountName: process.env.AUTH_ACCOUNTNAME,
+  oauthEndpoint: process.env.OAUTH_ENDPOINT,
+  authAccountName: process.env.AUTH_ACCOUNT_NAME,
+  authCompany: process.env.AUTH_COMPANY_NAME,
   authUserName: process.env.AUTH_USERNAME,
   authPassword: process.env.AUTH_PASSWORD,
 
   authGrantType: process.env.AUTH_GRANT_TYPE,
-  oauthEndpoint: process.env.OAUTH_ENDPOINT
+  oauthEndpoint: process.env.OAUTH_ENDPOINT,
+  baseUrl: process.env.BASE_URL,
 });
 
 // valid serviceCall DTO object