+++ src/core/client-config.model.ts
@@ -1,14 +1,18 @@
+
+export type OAuthGrantType = 'password' | 'client_credentials' | 'authorization_code';
 export type ClientConfig = {
   debug?: boolean | undefined;
   tokenCacheFilePath?: string | undefined;
 
   oauthEndpoint?: string | undefined;
+  
+  baseUrl: string;
 
   clientIdentifier: string;
   clientSecret: string;
   clientVersion: string;
 
-  authGrantType?: 'password' | 'client_credentials' | undefined;
+  authGrantType?: OAuthGrantType | undefined;
 
   authAccountName?: string | undefined;
   authUserName?: string | undefined;