+++ src/core/request-options.factory.ts
@@ -1,8 +1,8 @@
 import { v4 as uuid } from 'uuid';
 import { ALL_DTO_VERSIONS } from './all-dto-versions.constant';
-import { ClientConfig } from './client-config.model';
 import { DTOName } from './dto-name.model';
-import { OauthTokenResponse } from './oauth-token-response.model';
+import { OAuthTokenResponse } from './oauth/oauth-token-response.model';
+import { ClientConfig } from './client-config.model';
 
 export class RequestOptionsFactory {
 
@@ -14,14 +14,18 @@ export class RequestOptionsFactory {
     return Object.keys(o).map((key) => `${encodeURIComponent(key)}=${encodeURIComponent(o[key])}`).join('&');
   }
 
-  public static getDataApiUriFor(token: OauthTokenResponse, resourceName: DTOName, resourceId: string | null = null, externalId: string | null = null) {
+  private static getBaseUrl(config: Readonly<ClientConfig>) {
+    return config.baseUrl;
+  }
+
+  public static getDataApiUriFor(config: Readonly<ClientConfig>, resourceName: DTOName, resourceId: string | null = null, externalId: string | null = null) {
 
     const identifier = [
       (resourceId ? `/${resourceId}` : '').trim(),
       (externalId && !resourceId ? `/externalId/${externalId}` : '').trim()
     ].join('').trim();
 
-    return `${token.cluster_url}/api/data/v4/${resourceName}${identifier}`;
+    return `${this.getBaseUrl(config)}/api/data/v4/${resourceName}${identifier}`;
   }
 
   /**
@@ -43,17 +47,17 @@ export class RequestOptionsFactory {
       }).join(';');
   }
 
-  public static getRequestXHeaders(config: ClientConfig) {
+  public static getRequestXHeaders(config: Partial<{ clientIdentifier: string, clientVersion: string }>) {
     const requestId = uuid().replace(/\-/g, '');
     return {
-      'X-Client-Id': config.clientIdentifier,
-      'X-Client-Version': config.clientVersion,
+      'X-Client-Id': config.clientIdentifier || 'unknown',
+      'X-Client-Version': config.clientVersion || 'unknown',
       'X-Request-ID': requestId,
       'X-B3-TraceId': requestId
     }
   }
 
-  public static getRequestHeaders(token: OauthTokenResponse, config: ClientConfig) {
+  public static getRequestHeaders(token: { token_type: string, access_token: string }, config: { clientIdentifier: string, clientVersion: string }) {
     return {
       'Authorization': `${token.token_type} ${token.access_token}`,
       'Accept': 'application/json',
@@ -61,19 +65,26 @@ export class RequestOptionsFactory {
     }
   }
 
-  public static getRequestAccountQueryParams(token: OauthTokenResponse, config: ClientConfig) {
-    if (!token.companies || !token.companies.length) {
-      throw new Error('no compnay found on given account');
-    }
+  public static getRequestAccountQueryParams(token: OAuthTokenResponse,
+    config: Partial<{ authAccountName: string, authUserName: string, authCompany: string }>) {
 
-    const [selectedCompany] = token.companies;
+    const [firstCompany] = token.contentType === 'user'
+      && token.content.companies || [];
+
+    const company: string | undefined = config.authCompany
+      ? config.authCompany
+      : firstCompany?.name;
+
+    const user = config.authUserName
+      ? config.authUserName
+      : token.contentType === 'user'
+        ? token.content.user_name
+        : undefined;
 
     return {
+      user,
+      company,
       account: config.authAccountName,
-      user: config.authUserName,
-      company: config.authCompany
-        ? config.authCompany
-        : selectedCompany.name,
     }
   }
 }