+++ src/index.ts
@@ -1,5 +1,5 @@
 import { CoreAPIClient } from './core-api.client';
-import { CreateAction, UpdateAction, DeleteAction } from './core/batch-action.model';
+import { CreateAction, UpdateAction, DeleteAction } from './core/batch/batch-action.model';
 import { ALL_DTO_VERSIONS } from './core/all-dto-versions.constant';
 
 export {