+++ src/test/e2e/data-api.spec.ts
@@ -1,14 +1,14 @@
-import assert = require('assert');
-import { integrationTestConfig } from './integration-test.config';
-import { CoreAPIClient } from '../core-api.client';
-import { CreateAction, DeleteAction } from '..';
+import assert from 'assert';
+import { ClientConfigBuilder } from '../integration-test.config';
+import { CoreAPIClient } from '../../core-api.client';
+import { CreateAction, DeleteAction } from '../..';
 
 describe('DataApi', () => {
 
   let client: CoreAPIClient
 
   beforeEach(() => {
-    client = new CoreAPIClient(integrationTestConfig);
+    client = new CoreAPIClient(ClientConfigBuilder.getConfig('password'));
   });
 
   function prepareFixture() {
@@ -20,6 +20,9 @@ describe('DataApi', () => {
   const THE_ID = CoreAPIClient.createUUID();
   const THE_EXTERNAL_ID = `test-cleanup-${THE_ID}`;
 
+  let resolveWaiting: () => void;
+  const waitForObjectCreate = new Promise<void>((ok,) => (resolveWaiting = () => ok()));
+
   it('should throw on unknown DTOs', done => {
     client.getById('THIS_IS_NOT_A_DTO' as any, 'SOME-ID')
       .catch(error => {
@@ -28,7 +31,7 @@ describe('DataApi', () => {
       })
       .then(_ => done())
       .catch(error => done(error));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('POST serviceCall Id', done => {
 
@@ -42,13 +45,16 @@ describe('DataApi', () => {
       .then(fromDB => {
         const [{ serviceCall }] = fromDB.data;
         assert.strictEqual(serviceCall.id, THE_ID, 'POST');
-        assert.strictEqual(typeof (serviceCall as any as { createPerson: string }).createPerson, 'string');
+        assert.strictEqual(typeof (serviceCall as any as { lastChanged: string }).lastChanged, 'number');
+
+        resolveWaiting();
+
         return serviceCall;
       })
       .then(_ => done())
       .catch(e => done(e));
 
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('POST serviceCall ExternalId', done => {
     const MyExternalID = CoreAPIClient.createUUID();
@@ -61,13 +67,13 @@ describe('DataApi', () => {
       .then(fromDB => {
         const [{ serviceCall }] = fromDB.data;
         assert.strictEqual(serviceCall.externalId, MyExternalID, 'POST');
-        assert.strictEqual(typeof (serviceCall as any as { createPerson: { id: string } }).createPerson, 'object');
+        assert.strictEqual(typeof (serviceCall as any as { lastChanged: { id: string } }).lastChanged, 'number');
         return serviceCall;
       })
       .then(_ => done())
       .catch(e => done(e));
 
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('GET serviceCall by Id', done => {
 
@@ -80,7 +86,7 @@ describe('DataApi', () => {
       .then(_ => done())
       .catch(e => done(e));
 
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('GET serviceCall by ExternalId', done => {
     Promise.all([
@@ -88,16 +94,15 @@ describe('DataApi', () => {
       client.getByExternalId<{ id: string }>('ServiceCall', THE_EXTERNAL_ID)
     ])
       .then(([resp1, resp2]) => {
-        resp1.data[0].serviceCall.id;
         assert.strictEqual(resp1.data[0].serviceCall.id, THE_ID, 'POST');
-        assert.strictEqual(typeof (resp1.data[0].serviceCall as any as { createPerson: string }).createPerson, 'string');
+        assert.strictEqual(typeof (resp1.data[0].serviceCall as any as { lastChanged: string }).lastChanged, 'number');
         assert.strictEqual(resp2.data[0].serviceCall.id, resp1.data[0].serviceCall.id, 'POST ExternalId');
-        assert.strictEqual(typeof (resp2.data[0].serviceCall as any as { createPerson: { id: string } }).createPerson, 'object');
+        assert.strictEqual(typeof (resp2.data[0].serviceCall as any as { lastChanged: { id: string } }).lastChanged, 'number');
       })
       .then(_ => done())
       .catch(e => done(e));
 
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('PUT serviceCall by Id', done => {
     client.getById<{ id: string, subject: string }>('ServiceCall', THE_ID)
@@ -108,12 +113,12 @@ describe('DataApi', () => {
       .then(response => {
         const [{ serviceCall }] = response.data;
         assert.strictEqual(serviceCall.subject, 'subject-changed-with-PUT-ID', 'PUT by ID');
-        assert.strictEqual(typeof (serviceCall as any as { createPerson: string }).createPerson, 'string');
+        assert.strictEqual(typeof (serviceCall as any as { lastChanged: string }).lastChanged, 'number');
         return serviceCall;
       })
       .then(_ => done())
       .catch(e => done(e));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('PUT serviceCall by ExternalId', done => {
     client.getByExternalId<{ id: string, subject: string, lastChanged: number }>('ServiceCall', THE_EXTERNAL_ID)
@@ -124,12 +129,12 @@ describe('DataApi', () => {
       .then(response => {
         const [{ serviceCall }] = response.data;
         assert.strictEqual(serviceCall.subject, 'subject-changed-with-PUT-ExternalId', 'PUT by ExternalId');
-        assert.strictEqual(typeof (serviceCall as any as { createPerson: { id: string } }).createPerson, 'object');
+        assert.strictEqual(typeof (serviceCall as any as { lastChanged: { id: string } }).lastChanged, 'number');
         return serviceCall;
       })
       .then(_ => done())
       .catch(e => done(e));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('PATCH serviceCall by Id', done => {
     client.getById<{ id: string, lastChanged: number }>('ServiceCall', THE_ID)
@@ -144,12 +149,12 @@ describe('DataApi', () => {
       .then(response => {
         const [{ serviceCall }] = response.data;
         assert.strictEqual(serviceCall.subject, 'subject-changed-with-PATCH-ID', 'PATCH by ID');
-        assert.strictEqual(typeof (serviceCall as any as { createPerson: string }).createPerson, 'string');
+        assert.strictEqual(typeof (serviceCall as any as { lastChanged: string }).lastChanged, 'number');
         return serviceCall;
       })
       .then(_ => done())
       .catch(e => done(e));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('PATCH serviceCall by ExternalId', done => {
     client.getByExternalId<{ id: string, lastChanged: number, }>('ServiceCall', THE_EXTERNAL_ID)
@@ -164,12 +169,12 @@ describe('DataApi', () => {
       .then(response => {
         const [{ serviceCall }] = response.data;
         assert.strictEqual(serviceCall.subject, 'subject-changed-with-PATCH-ExternalId', 'PATCH by ExternalId');
-        assert.strictEqual(typeof (serviceCall as any as { createPerson: { id: string } }).createPerson, 'object');
+        assert.strictEqual(typeof (serviceCall as any as { lastChanged: { id: string } }).lastChanged, 'number');
         return serviceCall;
       })
       .then(_ => done())
       .catch(e => done(e));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('DELETE serviceCall by Id', done => {
     client.getById<{ id: string, lastChanged: number }>('ServiceCall', THE_ID)
@@ -181,7 +186,7 @@ describe('DataApi', () => {
       .then(fromDB => assert.strictEqual(fromDB.data.length, 0))
       .then(_ => done())
       .catch(e => done(e));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('DELETE serviceCall by externalId', done => {
     const id = CoreAPIClient.createUUID();
@@ -202,7 +207,7 @@ describe('DataApi', () => {
       .then(fromDB => assert.strictEqual(fromDB.data.length, 0))
       .then(_ => done())
       .catch(e => done(e));
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
   it('External Id E2E', done => {
 
@@ -239,6 +244,6 @@ describe('DataApi', () => {
       .then(_ => done())
       .catch(e => done(e));
 
-  }).timeout(5000);
+  }).timeout(ClientConfigBuilder.getTestTimeout());
 
 });